requirements.txt
websockets>=12.0
aiohttp>=3.9.0
pyarrow>=14.0.0
sortedcontainers>=2.4.0
asyncpg>=0.29.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
